requirements.txt
firebase-admin>=5.0.0
python-dotenv>=0.19.0
Flask>=2.0.0
Werkzeug>=2.0.0
